refactor: consolidate Daytona types into single module
Merge types_serialization, types_budget, types_context, types_recursive,
types_result into types.py. Reduces 5 files to 1. Update all imports.
